added animations and fixed sizes
